Works, layout not figured out

+++ foolscap_sitemap/crawl.sh
@@ -1,2 +1,5 @@
-rm foolscap.json
-scrapy crawl foolscap -t json -o foolscap.json -L WARN
+source ../bin/activate
+if [ -f foolscap.json ]; then
+    rm foolscap.json
+fi
+scrapy crawl foolscap -t json -o foolscap.json -L ${1:-WARN}
